1, fix power off/on issue in SmiDevice (add retry for TEST UTILITY) 2, disable ScsiCommand() throw exception, instead of returning false. 3, add power cycle feature to copy_compare_tester

+++ application/SmiVenderTest/SmiDevice/source/sm2236.cpp
@@ -347,7 +347,7 @@ void CSM2236::ReadSRAM(WORD ram_add, WORD bank, BYTE * buf)
 	cmd.add() = ram_add >> 8;
 	if ( (ram_add >= 0xC000) && (ram_add < 0xE000) && (bank != 0) )
 	{
-		cmd.bank(bank);
+		cmd.bank((BYTE)bank);
 	}
 	JCSIZE sec = 1;
 	VendorCommand(cmd, read, buf, sec);
